Moved _uniform_kernel to img_utils.py
