+++ neuroquery/meta_analysis.py
@@ -2,46 +2,9 @@
 
 import numpy as np
 import nibabel as nib
-
 from scipy.ndimage.filters import convolve
 
-
-def _uniform_kernel(r, n1=1, n2=1, n3=1):
-    """Build an uniform 3D kernel.
-
-    Parameters
-    ----------
-        r : int
-            Sphere radius >= 1.
-        n1 : int
-            Normalization factor on the 1st axis.
-        n2 : int
-            Normalization factor on the 2nd axis.
-        n3 : int
-            Normalization factor on the 3rd axis.
-
-    Returns:
-        array like
-            Array of shape (r//n1, r//n2, r//n3) storing the kernel.
-
-    """
-    A, B, C = r, r, r
-
-    a = int(A/abs(n1))
-    b = int(B/abs(n2))
-    c = int(C/abs(n3))
-
-    kernel = np.zeros((2*a+1, 2*b+1, 2*c+1))
-
-    i0, j0, k0 = a, b, c
-
-    for i in range(2*a+1):
-        for j in range(2*b+1):
-            for k in range(2*c+1):
-                if ((i-i0)/a)**2 + ((j-j0)/b)**2 + ((k-k0)/c)**2 <= 1:
-                    kernel[i, j, k] = 1
-
-    return kernel
+from neuroquery.img_utils import _uniform_kernel
 
 
 def _KDA_MKDA(peaks_arrs, affine, r=15, MKDA=False):
